requirements.txt
base58==2.1.1
bcrypt==4.0.1
blinker==1.9.0
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
Flask==3.1.0
Flask-Mail==0.10.0
future==0.18.3
gunicorn==23.0.0
idna==3.10
ipfshttpclient==0.7.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multiaddr==0.0.9
netaddr==1.3.0
netmiko==4.3.0
ntc_templates==4.0.1
packaging==24.2
paramiko==3.3.1
pycparser==2.21
PyNaCl==1.5.0
pyotp==2.9.0
pyserial==3.5
PyYAML==6.0.1
requests==2.32.3
scp==0.14.5
six==1.16.0
textfsm==1.1.3
urllib3==2.2.3
varint==1.0.2
Werkzeug==3.1.3
web3==6.5.0
python-dotenv==1.0.0
